Fix a few more pickle-related compatibility issues.
This converts one more case of a `pickle.loads()` call that had to be
converted to the `pickle_loads()` compatibility function, and updates
other call sites to not prematurely convert string types when feeding in
data to load.

Testing Done:
Unit tests pass for Python 2.7, 3.5, and 3.6.

Reviewed at https://reviews.reviewboard.org/r/10257/