Keep the injected on_delete out of the field signature.
To provide compatibility with Django 2.0+'s requirements for
`ForeignKey`, we recently pre-filled the `on_delete` attribute of the
field if not otherwise in the field signature. This was left in the
signature, causing problems during signature diffing. It's now removed
if it wasn't explicitly in there.

Testing Done:
Diffed a signature. `on_delete` didn't remain in.

Reviewed at https://reviews.reviewboard.org/r/10271/