Switch unit tests to the context manager form of assertRaisesMessage.
We have a lot of older unit tests that use the lambda form of
`assertRaisesMessage`, which is a bit more annoying to work with. This
was needed for older Python support. In preparation for some other
unit test changes, this change performs some cleanup on these calls,
switching to the context manager form in all calls.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/10293/