Add diffing support for the new signature classes.
This updates the diffing support to work with the new signature classes.
`Diff` now takes these instead of dictionaries. The capabilities and API
hasn't otherwise at this point, but in the future we'll be able to use
the new support to provide more fine-grained diffing so that only
subsets of the signature can be taken into account for evolutions.

Testing Done:
Unit tests pass for all supported versions of Python and Django.

Reviewed at https://reviews.reviewboard.org/r/10287/