Handle related_model specially in FieldSignature.
The original `FieldSignature` handled `related_model` as a special
argument, and a later version of that change got rid of that in favor of
a field attribute. The original idea was correct, as the equivalent
field attribute for `Field` classes required a model, whereas this
version represented a class path string.

This change restores this as its own attribute, and makes it easier to
construct a `FieldSignature` with a related model and field attributes
all at once. Diffing has been updated to factor this in as well.

Testing Done:
Tested along with other pending changes. Related model handling worked.

Reviewed at https://reviews.reviewboard.org/r/10295/