Move signature diffing into dedicated Signature.diff methods.
This introduces new `diff()` methods on signatures that focus on diffing
the content against another signature of the same type. These are used
by the main `Diff` class to perform the diffing (though in time that
class may disappear).

By splitting this out, the diffing logic becomes a lot more readable,
and we'll be able to use it for more fine-grained diffs and evolutions
down the road.

Testing Done:
All unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/10304/