Fix equality checking for signatures.
This change defines a `__ne__` for inequality checking, which is
necessary under Python 2.x (and the reason for some failed diffing that
was occurring). This just checks the inverse of `__eq__`.

It also fixes up most of the checks to ignore ordering in
`OrderedDicts` and most lists, so that two signatures can remain the
same even if, say, the order of fields changed.

Unit tests were added to ensure that all equality checks perform as
they're supposed to.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/10327/