Fix positional arguments for commands on modern Django.
This updates the `evolve`, `list-evolutions`, and `wipe-evolution`
commands to properly take positional arguments on modern versions of
Django. Previously, these were limited to keyword arguments, due to a
change in how argument parsing works now in Django. Positional arguments
have to be explicitly added to the arg parser, with a name of `args`.

The base compatibility class has also been updated to not break when
positional arguments are used on versions of Django using `optparse`.

Testing Done:
Ran these commands on Django 1.6 and 1.11 with positional arguments.

Reviewed at https://reviews.reviewboard.org/r/10546/