Improve scanning and normalization abilities for DatabaseState.
`DatabaseState` was originally built to help solve some tracking
problems for table indexes. We now need it to track model installation,
and going forward, we'll be using it to track table engines and other
details.

This change updates this class to fix up some small table lookup issues
(ensuring that we always normalize table names), to rename
`rescan_indexes()` to `rescan_tables()` (which is a more appropriate
name), and to add a utility function for determining if the database has
a table that backs a given model.

Future changes will update this to track additional information.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/10556/