Change Evolver task execution to prepare and execute in batches.
`Evolver` now groups tasks by their type, and works with them in batches
based on the type. That means all tasks of a type will be
prepared/executed at once.

It's now up to the task's class to handle the preparation and execution
of the tasks. This allows them to compute state before or after, and to
alter the state passed to each task.

When executing tasks, each batch of tasks of a given type are, by
default, executed in a transaction specific to the batch. Subclasses can
change the logic as needed for transaction handling.

Testing Done:
Unit tests pass on Django 1.7 and 1.11.

Reviewed at https://reviews.reviewboard.org/r/10566/