Add migration compatibility for Django 1.7 through 1.11.
The migration support recently added was built to work with the modern
functions and classes introduced in Django 1.10. Older versions had
different requirements in terms of the steps that needed to be
performed and the parameters that various functions took.

This change adds compatibility for Django 1.7 through 1.11. It also
appears to work in 2.0, but we are not currently guaranteeing any
support there.

Testing Done:
Unit tests pass on Django 1.6 through 1.11.

Reviewed at https://reviews.reviewboard.org/r/10568/