Rewrite the documentation for Django Evolution.
The documentation for Django Evolution was criminally obsolete,
referring to specific SVN revisions of Django that are needed, types of
supported fields that apparently could not be added, and limitations
that haven't existed for many, many years.

Given how big a release the upcoming 2.0 is, it's time (long past time)
to rewrite the docs.

The new docs provide an overview of the functionality of Django
Evolution, installation, and walkthroughs of how to work with
evolutions. There's a guide to all supported mutations and how to use
them, along with caveats to be aware of. There's also a FAQ with project
information, how evolutions and migrations interoperate, and some quirks
to expect while using evolutions.

There's command references for the `evolve`, `list-evolutions`, and
`wipe-evolution` commands, along with a codebase reference.

This will be expanded in the future as the project changes and grows,
but this is already a much better first step. The plan is to deploy
these docs on readthedocs.org.

An upcoming change will move the release notes over to these docs.

Testing Done:
Built the docs and read through them, checking for errors.

There are some build errors, but they're due to syntax issues in
source code that will be fixed separately.

Reviewed at https://reviews.reviewboard.org/r/10589/