Patch the SQLite backend to fix Alter Table breakages.
When running SQLite 2.26+ and Django 1.11 through 2.1.4, an Alter Table
would fail silently, causing problems further down the line. This was
due to some behavioral changes in SQLite when altering tables that
Django was not compatible with.

Django 2.1.4 turns on the `legacy_alter_table` flag when preparing to
edit schema. We now patch earlier versions of Django to do the same
thing.

Testing Done:
Tested in a Python environment with these versions and reproduced
the problem and the fix.

Reviewed at https://reviews.reviewboard.org/r/10869/