Add a MoveToMigrations evolution for all built-in Django apps.
We had several Django apps that were still being treated as owned by
Django Evolution, which meant that Django consistently warned about
unapplied migrations. While mostly harmless, this could become a problem
later.

This was caused by the lack of a `MoveToMigrations` built-in evolution
for each of these. We set it for a few apps, just the ones that we
needed to apply our own evolutions to at first, but didn't set it for
the remaining ones.

This change adds an evolution for this for every built-in Django app
that contains a `0001_initial.py` migration.

Testing Done:
Evolved the database successfully. Saw the migrations all apply.

Verified that Django no longer complained about missing migrations.

Reviewed at https://reviews.reviewboard.org/r/11075/