Add the beginnings of Django 2.2 support.
Django 2.2 removed some deprecated functionality, added new requirements
for multi-database unit tests, and again changed how table and index
names are normalized/hashed.

This change adds in compatibility for much of Django 2.2, minus some
work that still needs to be performed in SQLite (which many of the
recent changes works toward).

This does not implement the new `Meta.constraints` functionality, which
will need to be handled in another change.

Testing Done:
Most unit tests pass on Django 2.2 (and earlier versions). As this is not
a complete compatibility change, there's still work that needs to be done
before we can claim 2.2 support.

Reviewed at https://reviews.reviewboard.org/r/11086/